Moved MData and MultiSubs to their own File. Fixed for loops to be foreach loops. Altered postMulti and putMulti to accept datatype MData.
