+++ RedditSharp/RedditSharp.csproj
@@ -51,8 +51,10 @@
     <Compile Include="BotWebAgent.cs" />
     <Compile Include="Flairs\FlairTemplate.cs" />
     <Compile Include="Moderator Actions\ModActionType.cs" />
+    <Compile Include="Multi\MData.cs" />
     <Compile Include="Multi\Multi.cs" />
     <Compile Include="Multi\MultiData.cs" />
+    <Compile Include="Multi\MultiSubs.cs" />
     <Compile Include="Things\BannedUser.cs" />
     <Compile Include="Things\Contributor.cs" />
     <Compile Include="Things\ModAction.cs" />
